# ROX-13228: Performance benchmarking baselines for collection workflows