```GOROOT=/usr/local/opt/go/libexec #gosetup
GOPATH=/Users/oscarward/go #gosetup
/usr/local/opt/go/libexec/bin/go test -c -tags sql_integration -o /private/var/folders/44/6672qnyj1jg_8zgf2729kxh40000gn/T/GoLand/___BenchmarkCollections_in_github_com_stackrox_rox_central_resourcecollection_datastore.test github.com/stackrox/rox/central/resourcecollection/datastore #gosetup
/private/var/folders/44/6672qnyj1jg_8zgf2729kxh40000gn/T/GoLand/___BenchmarkCollections_in_github_com_stackrox_rox_central_resourcecollection_datastore.test -test.v -test.paniconexit0 -test.bench ^\QBenchmarkCollections\E$ -test.run ^$ -test.benchmem
goos: darwin
goarch: amd64
pkg: github.com/stackrox/rox/central/resourcecollection/datastore
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkCollections
BenchmarkCollections/DryRunAddCollection
BenchmarkCollections/DryRunAddCollection-8         	     313	   3866425 ns/op	 2569101 B/op	     413 allocs/op
BenchmarkCollections/DryRunUpdateCollection
BenchmarkCollections/DryRunUpdateCollection-8      	     274	   4439996 ns/op	 2573570 B/op	     478 allocs/op
BenchmarkCollections/UpdateCollection
BenchmarkCollections/UpdateCollection-8            	     178	   7334850 ns/op	 2907069 B/op	    3981 allocs/op
BenchmarkCollections/graphInit
BenchmarkCollections/graphInit-8                   	      12	 100315980 ns/op	11042746 B/op	  124287 allocs/op
BenchmarkCollections/AddCollection
BenchmarkCollections/AddCollection-8               	     130	   9770113 ns/op	 3311003 B/op	    7278 allocs/op
PASS

Process finished with the exit code 0
```